Add files via upload
If you don't want to tweak anything just the models, Pokemon ID NN.py, and the test images. The NN Gen file allows you to generate your own Neural Network to predict what generation a pokemon is from or really anything else for that matter. But be warned if you want to retrain it you will have to mass download a picture of every pokemon. In case you are wondering why I did this it was in an attempt to quantitatively show whether or not designs from certain generations were better or worse. I couldn't measure their quality directly so I figured if one generation was better than another it would at least be able to differentiate the two.